// packages/app/src/types/quake-api/games-summary-weapon.accuracy.interface.ts
export interface IGamesSummaryWeaponAccuracy {
    GAUNTLET: number;
    MACHINEGUN: number;
    MACHINEGUN_GRADE1: number;
    SHOTGUN: number;
    SHOTGUN_GRADE1: number;
    NAILGUN: number;
    NAILGUN_GRADE1: number;
    ROCKET_LAUNCHER: number;
    LIGHTNING_GUN: number;
    RAILGUN: number;
    LAGBOLT: number;
}